Add scabbards as a recognized demon force item
The demon force bot is no longer sensitive to quantities of demon force
items. If an item's uses have been exhausted, it will be blacklisted
from matching again for the rest of the bot session. It will also select
second and third effects when available.

Other changes:
- Matching for UI bar items is less stringent, and a more informative
error message will be logged on failure
- Wait longer after the initial window focus to ensure it is actually in
focus